requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
mysqlclient>=2.2,<3.0
python-decouple>=3.8,<4.0
Pillow>=11.0,<12.0
djangorestframework-simplejwt>=5.3,<6.0
